add docstrings of new functions
